fix: SCRUM-121 pricePerKb Integer에서 Double로 변경

+++ src/main/java/com/ixi_U/user/dto/response/ShowCurrentSubscribedResponse.java
@@ -9,7 +9,7 @@ public record ShowCurrentSubscribedResponse(
         String name,
         Integer mobileDataLimitMb,
         Integer monthlyPrice,
-        Integer pricePerKb,
+        Double pricePerKb,
         List<BundledBenefit> bundledBenefits,
         List<SingleBenefit> singleBenefits
 ) {
@@ -18,7 +18,7 @@ public record ShowCurrentSubscribedResponse(
             String name,
             Integer mobileDataLimitMb,
             Integer monthlyPrice,
-            Integer pricePerKb,
+            Double pricePerKb,
             List<BundledBenefit> bundledBenefits,
             List<SingleBenefit> singleBenefits
     ) {
